+++ src/main/java/com/palominolabs/metrics/guice/CountedInterceptor.java
@@ -8,11 +8,11 @@ import org.aopalliance.intercept.MethodInvocation;
 class CountedInterceptor implements MethodInterceptor {
 
     private final Counter counter;
-    private final Counted annotation;
+    private final boolean decrementAfterMethod;
 
     CountedInterceptor(Counter counter, Counted annotation) {
         this.counter = counter;
-        this.annotation = annotation;
+        decrementAfterMethod = !annotation.monotonic();
     }
 
     @Override
@@ -21,7 +21,7 @@ class CountedInterceptor implements MethodInterceptor {
         try {
             return invocation.proceed();
         } finally {
-            if (!annotation.monotonic()) {
+            if (decrementAfterMethod) {
                 counter.dec();
             }
         }