+++ src/test/java/com/palominolabs/metrics/guice/GaugeTest.java
@@ -11,8 +11,11 @@ import static com.codahale.metrics.MetricRegistry.name;
 import static com.palominolabs.metrics.guice.DefaultMetricNamer.GAUGE_SUFFIX;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertThat;
 
+@SuppressWarnings("unchecked")
 public class GaugeTest {
     private InstrumentedWithGauge instance;
     private MetricRegistry registry;
@@ -25,7 +28,6 @@ public class GaugeTest {
     }
 
     @Test
-    @SuppressWarnings("unchecked")
     public void aGaugeAnnotatedMethod() throws Exception {
         instance.doAThing();
 
@@ -41,7 +43,6 @@ public class GaugeTest {
     }
 
     @Test
-    @SuppressWarnings("unchecked")
     public void aGaugeAnnotatedMethodWithDefaultName() throws Exception {
         instance.doAnotherThing();
 
@@ -58,7 +59,6 @@ public class GaugeTest {
     }
 
     @Test
-    @SuppressWarnings("unchecked")
     public void aGaugeAnnotatedMethodWithAbsoluteName() throws Exception {
         instance.doAThingWithAbsoluteName();
 
@@ -71,5 +71,29 @@ public class GaugeTest {
         assertThat("Guice creates a gauge with the given value",
             ((Gauge<String>) metric).getValue(),
             is("anotherThingWithAbsoluteName"));
+    }
+
+    @Test
+    public void aGaugeInSuperclass() throws Exception {
+        final Gauge<?> metric = registry.getGauges().get(name("gaugeParent"));
+
+        assertNotNull(metric);
+        assertEquals("gaugeParent", metric.getValue());
+    }
+
+    @Test
+    public void aPrivateGaugeInSuperclass() throws Exception {
+        final Gauge<?> metric = registry.getGauges().get(name("gaugeParentPrivate"));
+
+        assertNotNull(metric);
+        assertEquals("gaugeParentPrivate", metric.getValue());
+    }
+
+    @Test
+    public void aPrivateGauge() throws Exception {
+        final Gauge<?> metric = registry.getGauges().get(name(InstrumentedWithGauge.class, "gaugePrivate"));
+
+        assertNotNull(metric);
+        assertEquals("gaugePrivate", metric.getValue());
     }
 }