+++ src/test/java/com/palominolabs/metrics/guice/GenericThing.java
@@ -1,0 +1,8 @@
+package com.palominolabs.metrics.guice;
+
+class GenericThing<T> {
+
+    void doThing(T t) {
+
+    }
+}