+++ src/test/java/com/palominolabs/metrics/guice/InstrumentedWithCounterParent.java
@@ -1,0 +1,11 @@
+package com.palominolabs.metrics.guice;
+
+import com.codahale.metrics.annotation.Counted;
+
+public class InstrumentedWithCounterParent {
+
+    @Counted(name = "counterParent", absolute = true)
+    public String counterParent() {
+        return "counterParent";
+    }
+}