+++ src/test/java/com/palominolabs/metrics/guice/InstrumentedWithGauge.java
@@ -2,7 +2,7 @@ package com.palominolabs.metrics.guice;
 
 import com.codahale.metrics.annotation.Gauge;
 
-public class InstrumentedWithGauge {
+public class InstrumentedWithGauge extends InstrumentedWithGaugeParent {
     @Gauge(name = "things")
     public String doAThing() {
         return "poop";
@@ -16,5 +16,10 @@ public class InstrumentedWithGauge {
     @Gauge(name = "absoluteName", absolute = true)
     public String doAThingWithAbsoluteName() {
         return "anotherThingWithAbsoluteName";
+    }
+
+    @Gauge(name = "gaugePrivate")
+    private String gaugePrivate() {
+        return "gaugePrivate";
     }
 }