+++ RELEASE-NOTES.md
@@ -1,3 +1,7 @@
+- TBD
+    - License under COIL
+    - Add ability to have non-public @Gauge methods anywhere in the type hierarchy of an injected type
+    - Updated dependencies: Metrics 3.1.2, SLF4J 1.7.12, Guice 4.0
 - 3.1.3
     - Add support for `@Counted`
     - Move metric name creation into `MetricNamer` for easy customization